help: sopel_help.providers setup entry point
This is the big deal of this plugin: using a setup entry point to
configure the list of available help providers.

So far are included:

* base (default): publish directly to the user
* clbin: publish to clbin.com (using HTTPS)
* 0x0: publish to 0x0.st (using HTTPS)
* hastebin: publish to hastebin.com (using HTTPS)
* termbin: publish to termbin.com (using socket)
* ubuntu: publish to pastebin.ubuntu.com (using HTTPS)

The provider is selected at the plugin's setup phase, by using a Manager
that exposes the list of available providers.

A configuration section is defined, which is backward compatible with
the Sopel's built-in help plugin.

All providers and the manager are fully tested.