managers, meta: migrate off of `pkg_resources` to `importlib_metadata`
This fixes installing `sopel` (which depends on this package) from
scratch on a new environment that does not ship with `pkg_resources` as
part of the Python stdlib (py3.12+) and also omits `setuptools` (some
distros, fresh venvs, probably others).

Includes slight cleanup of comments, log messages, and imports, but I
think not enough to be worth a separate commit or pull request.